backend/requirements.txt
fastapi
uvicorn
python-multipart
PyPDF2
chromadb
sentence-transformers
openai 
python-dotenv
